Move all struct  to struct folder
