+++ src/structures/bytearray.rs
