+++ src/utils.rs
@@ -1,7 +1,10 @@
 //! Utils
 //!
 //! Various utils functions defined for the KEM anf PKE algorithms
-use crate::{hash, polyvec::structures::FiniteField, ByteArray, Poly3329, F3329};
+use crate::{
+    hash,
+    structures::{algebraics::FiniteField, ByteArray, Poly3329, F3329},
+};
 
 /// Receives as input a byte stream B=(b0; b1; b2;...) and computes the NTT-representation a' = a'_0 + a'_0X + ... + a'_n-1X^(n-1) in R_q of a in R_q
 /// Algorithm 1 p. 7
